Pridávanie transakcii pomocou tlačidla add transaction

+++ app/src/main/java/com/meloge/uceniedatabazafinal/MainActivity.kt
@@ -11,9 +11,13 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.lifecycleScope
 import com.meloge.uceniedatabazafinal.data.model.Transaction
+import com.meloge.uceniedatabazafinal.ui.screen.AddTransactionScreen
 import com.meloge.uceniedatabazafinal.ui.screen.TransactionListScreen
 import com.meloge.uceniedatabazafinal.ui.theme.UcenieDatabazaFinalTheme
 import com.meloge.uceniedatabazafinal.ui.viewmodel.TransactionViewModel
@@ -28,7 +32,7 @@ class MainActivity : ComponentActivity() {
 
         lifecycleScope.launch {
             //transactionViewModel.deleteAllTransactions()
-            transactionViewModel.insertTransaction(Transaction(amount = 10.0, description = "Test"))
+            //transactionViewModel.insertTransaction(Transaction(amount = 10.0, description = "Test"))
         }
 
         setContent {
@@ -38,15 +42,34 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background,
                 ) {
+                    var showAddTransactionScreen by remember { mutableStateOf(false) }
                     val transactions by transactionViewModel.transactions.collectAsState()
-                    TransactionListScreen(
-                        transactions = transactions,
-                        onDeleteAllTransactions = {
+
+                    if (showAddTransactionScreen) {
+                        AddTransactionScreen(onAddTransaction = { amount, description ->
                             lifecycleScope.launch {
-                                transactionViewModel.deleteAllTransactions()
+                                transactionViewModel.insertTransaction(
+                                    Transaction(
+                                        amount = amount,
+                                        description = description
+                                    )
+                                )
                             }
-                        }
-                    )
+                            showAddTransactionScreen = false
+                        })
+                    } else {
+                        TransactionListScreen(
+                            transactions = transactions,
+                            onDeleteAllTransactions = {
+                                lifecycleScope.launch {
+                                    transactionViewModel.deleteAllTransactions()
+                                }
+                            },
+                            onNavigateToAddTransaction = {
+                                showAddTransactionScreen = true
+                            }
+                        )
+                    }
                 }
             }
         }
